Revert ":sparkles:  New pool list (#48)"
This reverts commit 610b3453159d33efb59b56d5a843c89d92f2a22d.
