+++ views/create-pool/illustration/coins-illustration.tsx
