+++ views/create-pool/illustration/index.tsx
