+++ views/create-pool/pool-status/index.tsx
