+++ views/create-pool/pool-status/pool-status.types.ts
