+++ views/pools/pool-form/component/receive-section/receive-section.types.ts
