--- views/pools/pool-list/my-position-section.tsx
@@ -1,12 +1,0 @@
-import { Box } from '@interest-protocol/ui-kit';
-import { FC } from 'react';
-
-const MyPosition: FC = () => {
-  return (
-    <Box p="xs" my="1.5rem" gap="0.5rem">
-      My Position
-    </Box>
-  );
-};
-
-export default MyPosition;
