--- views/pools/pool-list/pool-list.data.tsx
@@ -1,351 +1,0 @@
-import { Box } from '@interest-protocol/ui-kit';
-
-import { Routes, RoutesEnum } from '@/constants';
-import { DefaultTokenSVG, IPXRoundedSVG } from '@/svg';
-
-import { PoolCardProps } from './components/pools-card/pool-card.types';
-
-export const PoolListData: ReadonlyArray<PoolCardProps> = [
-  {
-    dexInfo: {
-      name: 'Interest',
-      url: Routes[RoutesEnum.PoolDetails],
-      Logo: (
-        <Box
-          p="S"
-          width="1.3rem"
-          height="1.3rem"
-          display="flex"
-          color="lowestContainer"
-          alignItems="center"
-          borderRadius="full"
-          justifyContent="center"
-          bg="transparent"
-        >
-          <IPXRoundedSVG
-            maxWidth="1.125rem"
-            maxHeight="1.125rem"
-            width="100%"
-          />
-        </Box>
-      ),
-    },
-    pairInfo: {
-      apr: '333.45',
-      tokenList: [
-        {
-          symbol: 'USDT',
-          Icon: DefaultTokenSVG,
-        },
-        {
-          symbol: 'USDT',
-          Icon: DefaultTokenSVG,
-        },
-      ],
-    },
-    tradeInfo: {
-      lines: [
-        {
-          description: 'Fee',
-          amount: '000%',
-          tooltipInfo: 'Fee per transaction',
-        },
-        {
-          description: 'Liquidity',
-          amount: '$234.566',
-          tooltipInfo: 'Liquidity',
-        },
-        {
-          description: 'Volume (24)',
-          amount: '$234.88',
-          tooltipInfo: 'Volume',
-        },
-      ],
-    },
-  },
-  {
-    dexInfo: {
-      name: 'Interest',
-      url: Routes[RoutesEnum.PoolDetails],
-      Logo: (
-        <Box
-          p="S"
-          width="1.3rem"
-          height="1.3rem"
-          display="flex"
-          color="lowestContainer"
-          alignItems="center"
-          borderRadius="full"
-          justifyContent="center"
-          bg="transparent"
-        >
-          <IPXRoundedSVG
-            maxWidth="1.125rem"
-            maxHeight="1.125rem"
-            width="100%"
-          />
-        </Box>
-      ),
-    },
-    pairInfo: {
-      apr: '333.45',
-      tokenList: [
-        {
-          symbol: 'USDT',
-          Icon: DefaultTokenSVG,
-        },
-        {
-          symbol: 'USDT',
-          Icon: DefaultTokenSVG,
-        },
-      ],
-    },
-    tradeInfo: {
-      lines: [
-        {
-          description: 'Fee',
-          amount: '000%',
-          tooltipInfo: 'Fee per transaction',
-        },
-        {
-          description: 'Liquidity',
-          amount: '$234.566',
-          tooltipInfo: 'Liquidity',
-        },
-        {
-          description: 'Volume (24)',
-          amount: '$234.88',
-          tooltipInfo: 'Volume',
-        },
-      ],
-    },
-  },
-  {
-    dexInfo: {
-      name: 'Interest',
-      url: Routes[RoutesEnum.PoolDetails],
-      Logo: (
-        <Box
-          p="S"
-          width="1.3rem"
-          height="1.3rem"
-          display="flex"
-          color="lowestContainer"
-          alignItems="center"
-          borderRadius="full"
-          justifyContent="center"
-          bg="transparent"
-        >
-          <IPXRoundedSVG
-            maxWidth="1.125rem"
-            maxHeight="1.125rem"
-            width="100%"
-          />
-        </Box>
-      ),
-    },
-    pairInfo: {
-      apr: '333.45',
-      tokenList: [
-        {
-          symbol: 'USDT',
-          Icon: DefaultTokenSVG,
-        },
-        {
-          symbol: 'USDT',
-          Icon: DefaultTokenSVG,
-        },
-      ],
-    },
-    tradeInfo: {
-      lines: [
-        {
-          description: 'Fee',
-          amount: '000%',
-          tooltipInfo: 'Fee per transaction',
-        },
-        {
-          description: 'Liquidity',
-          amount: '$234.566',
-          tooltipInfo: 'Liquidity',
-        },
-        {
-          description: 'Volume (24)',
-          amount: '$234.88',
-          tooltipInfo: 'Volume',
-        },
-      ],
-    },
-  },
-  {
-    dexInfo: {
-      name: 'Interest',
-      url: Routes[RoutesEnum.PoolDetails],
-      Logo: (
-        <Box
-          p="S"
-          width="1.3rem"
-          height="1.3rem"
-          display="flex"
-          color="lowestContainer"
-          alignItems="center"
-          borderRadius="full"
-          justifyContent="center"
-          bg="transparent"
-        >
-          <IPXRoundedSVG
-            maxWidth="1.125rem"
-            maxHeight="1.125rem"
-            width="100%"
-          />
-        </Box>
-      ),
-    },
-    pairInfo: {
-      apr: '333.45',
-      tokenList: [
-        {
-          symbol: 'USDT',
-          Icon: DefaultTokenSVG,
-        },
-        {
-          symbol: 'USDT',
-          Icon: DefaultTokenSVG,
-        },
-      ],
-    },
-    tradeInfo: {
-      lines: [
-        {
-          description: 'Fee',
-          amount: '000%',
-          tooltipInfo: 'Fee per transaction',
-        },
-        {
-          description: 'Liquidity',
-          amount: '$234.566',
-          tooltipInfo: 'Liquidity',
-        },
-        {
-          description: 'Volume (24)',
-          amount: '$234.88',
-          tooltipInfo: 'Volume',
-        },
-      ],
-    },
-  },
-  {
-    dexInfo: {
-      name: 'Interest',
-      url: Routes[RoutesEnum.PoolDetails],
-      Logo: (
-        <Box
-          p="S"
-          width="1.3rem"
-          height="1.3rem"
-          display="flex"
-          color="lowestContainer"
-          alignItems="center"
-          borderRadius="full"
-          justifyContent="center"
-          bg="transparent"
-        >
-          <IPXRoundedSVG
-            maxWidth="1.125rem"
-            maxHeight="1.125rem"
-            width="100%"
-          />
-        </Box>
-      ),
-    },
-    pairInfo: {
-      apr: '333.45',
-      tokenList: [
-        {
-          symbol: 'USDT',
-          Icon: DefaultTokenSVG,
-        },
-        {
-          symbol: 'USDT',
-          Icon: DefaultTokenSVG,
-        },
-      ],
-    },
-    tradeInfo: {
-      lines: [
-        {
-          description: 'Fee',
-          amount: '000%',
-          tooltipInfo: 'Fee per transaction',
-        },
-        {
-          description: 'Liquidity',
-          amount: '$234.566',
-          tooltipInfo: 'Liquidity',
-        },
-        {
-          description: 'Volume (24)',
-          amount: '$234.88',
-          tooltipInfo: 'Volume',
-        },
-      ],
-    },
-  },
-  {
-    dexInfo: {
-      name: 'Interest',
-      url: Routes[RoutesEnum.PoolDetails],
-      Logo: (
-        <Box
-          p="S"
-          width="1.3rem"
-          height="1.3rem"
-          display="flex"
-          color="lowestContainer"
-          alignItems="center"
-          borderRadius="full"
-          justifyContent="center"
-          bg="transparent"
-        >
-          <IPXRoundedSVG
-            maxWidth="1.125rem"
-            maxHeight="1.125rem"
-            width="100%"
-          />
-        </Box>
-      ),
-    },
-    pairInfo: {
-      apr: '333.45',
-      tokenList: [
-        {
-          symbol: 'USDT',
-          Icon: DefaultTokenSVG,
-        },
-        {
-          symbol: 'USDT',
-          Icon: DefaultTokenSVG,
-        },
-      ],
-    },
-    tradeInfo: {
-      lines: [
-        {
-          description: 'Fee',
-          amount: '000%',
-          tooltipInfo: 'Fee per transaction',
-        },
-        {
-          description: 'Liquidity',
-          amount: '$234.566',
-          tooltipInfo: 'Liquidity',
-        },
-        {
-          description: 'Volume (24)',
-          amount: '$234.88',
-          tooltipInfo: 'Volume',
-        },
-      ],
-    },
-  },
-];
