--- views/pools/pool-list/pool-list.types.ts
@@ -1,4 +1,0 @@
-export enum PoolTypeEnum {
-  Pool,
-  MyPosition,
-}
