+++ views/pools/pool-transaction/information.tsx
