+++ views/pools/pool-transaction/statistics.tsx
