+++ views/pools/pools-card/pool-card.types.ts
@@ -1,6 +1,6 @@
 import { FC, ReactNode } from 'react';
 
-import { SVGProps } from '@/components/svg/svg.types';
+import { SVGProps } from '../../../components/svg/svg.types';
 
 export interface TokenProps {
   Icon: FC<SVGProps>;
@@ -19,18 +19,11 @@ export interface PoolCardTokenInfoProps {
 }
 
 export interface PoolCardLineProps {
-  amount: string;
-  lastLine?: boolean;
   description: string;
+  amount: string;
   tooltipInfo: string;
 }
 
 export interface PoolTradeInfoProps {
   lines: ReadonlyArray<PoolCardLineProps>;
-}
-
-export interface PoolCardProps {
-  dexInfo: PoolCardHeaderProps;
-  pairInfo: PoolCardTokenInfoProps;
-  tradeInfo: PoolTradeInfoProps;
 }
