+++ views/pools/pools-card/token-info.tsx
